% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUCtrap.R
\name{AUCtrap}
\alias{AUCtrap}
\title{Calculate AUC Using Trapezoidal Rule}
\usage{
AUCtrap(x, y, method = c("AUC", "iAUC", "netAUC", "minAUC"))
}
\arguments{
\item{x}{numeric vector giving the measurement times (x-values)}

\item{y}{numeric vector with the measurements corresponding to \code{x}}

\item{method}{character value defining the calculation method. Defaults
to "AUC", which is the total area under the curve}
}
\value{
numeric value with the calculated area
}
\description{
\code{AUCtrap} implements several methods for calculating the area
under a piecewise linear function defined by longitudinally measured
values of a variable.
}
\details{
The implemented methods are total AUC ("AUC"), incremental AUC measured
from the first as baseline ("iAUC"), net AUC ("netAUC"), and AUC from the
minimum as baseline ("minAUC").
}
\examples{
AUCtrap(0:3, c(3,4,1,2))
}
